Split Wms.Goods.Properties tests in their own module
