+++ anyblok_wms_base/core/goods/tests/test_properties.py
@@ -1,0 +1,125 @@
+# -*- coding: utf-8 -*-
+# This file is a part of the AnyBlok / WMS Base project
+#
+#    Copyright (C) 2018 Georges Racinet <gracinet@anybox.fr>
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v. 2.0. If a copy of the MPL was not distributed with this file,You can
+# obtain one at http://mozilla.org/MPL/2.0/.
+# -*- coding: utf-8 -*-
+# This file is a part of the AnyBlok / WMS Base project
+#
+#    Copyright (C) 2018 Georges Racinet <gracinet@anybox.fr>
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v. 2.0. If a copy of the MPL was not distributed with this file,You can
+# obtain one at http://mozilla.org/MPL/2.0/.class TestGoodsProperties(BlokTestCase):
+"""Tests of Wms.Goods.Properties.
+
+This is not about tests of Wms.Goods involving properties
+(for these, see test_goods).
+"""
+
+from anyblok.tests.testcase import BlokTestCase
+
+
+class TestGoodsProperties(BlokTestCase):
+    def setUp(self):
+        self.Props = self.registry.Wms.Goods.Properties
+
+    def test_get_set(self):
+        props = self.Props.insert()
+
+        # at this stage, props.flexible is None, but get() and [] must behave
+        # as if it were an empty dict
+        with self.assertRaises(KeyError) as arc:
+            props['foo']
+        self.assertEqual(arc.exception.args, ('foo', ))
+        self.assertIsNone(props.get('foo'))
+        self.assertEqual(props.get('foo', 4), 4)
+
+        props['foo'] = 3
+        props[123] = 'bar'
+        props['batch'] = 1
+
+        self.assertEqual(props['foo'], 3)
+        self.assertEqual(props.get('foo'), 3)
+
+        self.assertEqual(props[123], 'bar')
+        self.assertEqual(props.get(123), 'bar')
+
+        self.assertEqual(props['batch'], 1)
+        self.assertEqual(props.get('batch'), 1)
+
+        self.assertIsNone(props.get('missing'))
+        self.assertEqual(props.get('missing', 1.2), 1.2)
+
+        with self.assertRaises(TypeError):
+            self.assertEqual(props.get('a', 'default', 'extra'))
+
+    def test_update(self):
+        props = self.Props()
+
+        with self.assertRaises(TypeError):
+            # two positional arguments
+            props.update(dict(a=1), dict(b=2))
+
+        # a dict positional arg
+        props.update(dict(batch=1))
+        # an iterable of pairs
+        props.update([('x', 2), ('y', 3)])
+        # keyword arguments
+        props.update(z=5, foo='bar')
+
+        self.assertEqual(props.as_dict(),
+                         dict(batch=1, x=2, y=3, z=5, foo='bar'))
+
+        # both positional and keywords arguments (yes it works with dicts)
+        # also, overwritting existing properties
+        props.update(dict(batch=3), z=['a', 'b'], foo='spam')
+        self.assertEqual(props.as_dict(),
+                         dict(batch=3, x=2, y=3, z=['a', 'b'], foo='spam'))
+
+    def test_create(self):
+        props = self.Props.create(batch='abcd',
+                                  serial=1234, expiry='2018-03-01')
+        self.assertEqual(props.to_dict(),
+                         dict(batch='abcd',
+                              id=props.id,
+                              flexible=dict(serial=1234, expiry='2018-03-01')))
+
+        self.assertIsNone(self.Props.create())
+
+    def test_reserved(self):
+        with self.assertRaises(ValueError):
+            self.Props.create(batch='abcd', flexible=True)
+
+    def test_as_dict(self):
+        props = self.Props.create(batch='abcd', history=['a'], serial=2345)
+        as_dict = props.as_dict()
+        self.assertEqual(as_dict,
+                         dict(batch='abcd', history=['a'], serial=2345))
+
+        # mutability issues
+        as_dict['history'].append('b')
+        self.assertEqual(props.get('history'), ['a'])
+
+    def test_contains(self):
+        props = self.Props.insert(batch='abc')
+        self.assertTrue('batch' in props)
+        self.assertFalse('x' in props)  # props.flexible is None
+        props['x'] = 1
+        self.assertTrue('x' in props)
+        self.assertFalse('y' in props)  # props.flexible isn't None
+
+    def test_duplicate(self):
+        props = self.Props.create(batch='abcd', history=['a'], serial=2345)
+        dup = props.duplicate()
+        self.assertNotEqual(dup.id, props.id)
+        self.assertEqual(dup.as_dict(),
+                         dict(batch='abcd', history=['a'], serial=2345))
+
+        # niceties with mutability
+        dup.get('history').append('b')
+        self.assertEqual(dup.get('history'), ['a', 'b'])
+        self.assertEqual(props.get('history'), ['a'])
